Move 'helper' to utils
