+++ commands/deploy_sources.py
@@ -1,6 +1,6 @@
 from os import path
 import boto3
-import helper
+from utils import helper
 from utils.logger import log
 
 s3_client = boto3.client('s3')
